+++ KoS.Apps.SharePoint.SmartCAML/KoS.Apps.SharePoint.SmartCAML.Editor/Utils/MyCommands.cs
@@ -11,6 +11,8 @@ namespace KoS.Apps.SharePoint.SmartCAML.Editor.Utils
         public static RoutedCommand CloseQueryTab = new RoutedCommand();
         public static RoutedUICommand HideColumn = new RoutedUICommand();
         public static RoutedUICommand HideAllColumn = new RoutedUICommand();
+        public static RoutedUICommand HideAllHiddenColumns = new RoutedUICommand();
+        public static RoutedUICommand HideAllReadonlyColumns = new RoutedUICommand();
         public static RoutedUICommand HideUnPinedColumn = new RoutedUICommand();
         public static RoutedUICommand UnhideColumn = new RoutedUICommand();
         public static RoutedUICommand UnhideAllColumn = new RoutedUICommand();